Refactor backend ImageManager into utility ImageHandler; Implement Firebase Storage abstraction in ProfileActivity
